set and save data thru save button on toolbar, set handler methods

+++ app/src/main/java/com/example/androidexampleapp/model/DataStore.kt
@@ -32,7 +32,7 @@ object DataStore {
         val id = database?.insertCar(car) ?: return
 
         if (id > 0) {
-            car.id = id.toInt()
+            car.id = id
             cars.add(car)
 
             Log.d("SQLite!", "Car successfully added")
